+++ src/extension-test/unit/formatting-test.ts
@@ -2,57 +2,82 @@ import * as expect from 'expect';
 import { formatIndex } from '../../calva-fmt/src/format-index';
 import { backspaceOnWhitespace } from '../../cursor-doc/backspace-on-whitespace';
 import * as indent from '../../cursor-doc/indent';
-
 import { docFromTextNotation, textAndSelection } from './common/text-notation';
 
 describe('formatter, indenter and paredit comparison', () => {
   const configs = [
-    mkConfig(
-      {
-        '/\\S+/': [['inner', 0]],
-      },
-      '{:indents {#"\\S+" [[:inner 0]]}}'
-    ),
-    mkConfig(
-      {
-        '/\\S+/': [['inner', 1]],
-      },
-      '{:indents {#"\\S+" [[:inner 1]]}}'
-    ),
+    mkConfig({
+      '#"\\S+"': [['inner', 0]],
+    }),
+    mkConfig({
+      '#"\\S+"': [['block', 0]],
+    }),
+    mkConfig({
+      '#"\\S+"': [['inner', 1]],
+    }),
   ];
 
-  configs.forEach((config) => {
-    it(`correctly formats 'and' form with formatter, indenter and paredit using config: ${config['cljfmt-options-string']}`, () => {
-      const formatterIndent = getFormatterIndent('(and x\ny)', 7, config);
-      const indenterIndent = getIndenterIndent('(and x\n|y)', config);
-      const pareditIndent = getPareditIndent('(and x\n\n  |y)', 10, config);
+  describe('indents `and` form the same with formatter, indenter, and paredit', () => {
+    configs.forEach((config) => {
+      it(`indents 'and' form the same with formatter, indenter, and paredit, using config: ${config['cljfmt-options-string']}`, () => {
+        const formatterIndent = getFormatterIndent('(and x\n|y)', config);
+        const indenterIndent = getIndenterIndent('(and x\n|y)', config);
+        const pareditIndent = getPareditIndent('(and x\n\n  |y)', config);
+        expect(formatterIndent).toEqual(indenterIndent);
+        expect(formatterIndent).toEqual(pareditIndent);
+      });
+    });
+  });
 
-      expect(formatterIndent).toEqual(indenterIndent);
-      expect(formatterIndent).toEqual(pareditIndent);
+  describe('indents keyword form the same with formatter, indenter and paredit', () => {
+    configs.forEach((config) => {
+      it(`indents keyword form the same with formatter, indenter, and paredit, using config: ${config['cljfmt-options-string']}`, () => {
+        const formatterIndent = getFormatterIndent('(:kw x\n|y)', config);
+        const indenterIndent = getIndenterIndent('(:kw x\n|y)', config);
+        const pareditIndent = getPareditIndent('(:kw x\n\n  |y)', config);
+        expect(formatterIndent).toEqual(indenterIndent);
+        expect(formatterIndent).toEqual(pareditIndent);
+      });
     });
   });
 });
 
 function getIndenterIndent(form: string, config: ReturnType<typeof mkConfig>) {
-  const indenterDoc = docFromTextNotation(form);
-  return indent.getIndent(indenterDoc.model, textAndSelection(indenterDoc)[1][0], config);
+  const doc = docFromTextNotation(form);
+  const p = textAndSelection(doc)[1][0];
+  return indent.getIndent(doc.model, p, config);
 }
 
-function getFormatterIndent(form: string, cursor: number, config: ReturnType<typeof mkConfig>) {
-  const formatterResult = formatIndex(form, [0, form.length], cursor, '\n', false, config);
+function getFormatterIndent(notation: string, config: ReturnType<typeof mkConfig>) {
+  const doc = docFromTextNotation(notation);
+  const form = textAndSelection(doc)[0];
+  const p = textAndSelection(doc)[1][0];
+  const formatterResult = formatIndex(form, [0, notation.length], p, '\n', false, config);
   return formatterResult['new-index'] - formatterResult.idx;
 }
 
-function getPareditIndent(form: string, cursor: number, config: ReturnType<typeof mkConfig>) {
-  const pareditDoc = docFromTextNotation(form);
-  return backspaceOnWhitespace(pareditDoc, pareditDoc.getTokenCursor(cursor), config).indent;
+function getPareditIndent(notation: string, config: ReturnType<typeof mkConfig>) {
+  const doc = docFromTextNotation(notation);
+  const p = textAndSelection(doc)[1][0];
+  return backspaceOnWhitespace(doc, doc.getTokenCursor(p), config).indent;
 }
 
-function mkConfig(rules: indent.IndentRules, rulesString: string) {
+function mkConfig(rules: indent.IndentRules) {
+  const cljRules = jsRulesToCljsRulesString(rules);
   return {
-    'cljfmt-options-string': rulesString,
+    'cljfmt-options-string': `{:indents {${cljRules}}}`,
     'cljfmt-options': {
       indents: rules,
     },
   };
+}
+
+function jsRulesToCljsRulesString(rules: indent.IndentRules) {
+  return Object.entries(rules).reduce((acc, [k, v], i) => {
+    return acc + `${k} ${jsRuleToCljRuleString(v)} `;
+  }, '');
+}
+
+function jsRuleToCljRuleString(value: indent.IndentRule[]) {
+  return '[' + value.map((a) => `[:${a[0]} ${a[1]}]`).join('') + ']';
 }