+++ CHANGELOG.md
@@ -4,6 +4,12 @@ Changes to Calva.
 
 ## [Unreleased]
 
+## [2.0.331] - 2023-02-05
+
+- Bump npm deps [jszip](https://github.com/BetterThanTomorrow/calva/pull/2056), [http-cache-semantics](https://github.com/BetterThanTomorrow/calva/pull/2059)
+- Fix: [Indenter and formatter do not agree on keyword in function position when regex indent rules are involved](https://github.com/BetterThanTomorrow/calva/issues/2044)
+- Fix: [Missing required argument for "-M ALIASES"](https://github.com/BetterThanTomorrow/calva/issues/2039)
+
 ## [2.0.330] - 2023-02-03
 
 - Fix: [Clojure-lsp does not automatically start when using VSCode on Windows in Calva version 2.0.329](https://github.com/BetterThanTomorrow/calva/issues/2054)